base cards done
